requirements.txt
anyio==3.7.1
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.1.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dateparser==1.1.8
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
fastapi==0.99.1
fonttools==4.40.0
h11==0.14.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
murmurhash==1.0.9
numpy==1.25.0
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.0.0
preshed==3.0.8
pycparser==2.21
pydantic==1.10.11
pyparsing==3.0.9
PyPDF2==2.12.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.27.0
tabula-py==2.7.0
thinc==8.1.10
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
uvicorn==0.22.0
Wand==0.6.11
wasabi==1.1.2
